requirements.txt
anthropic>=0.84.0
httpx>=0.28.1
requests>=2.31.0
pydantic>=2.12.5
numpy
pandas
beautifulsoup4>=4.12
lxml>=5.0
tenacity>=9.0
python-dotenv>=1.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
